Fix implicit conversion error.

+++ src/wasm/wasm-module.h
@@ -65,7 +65,7 @@ struct WasmModule {
 
   // Checks the given offset range is contained within the module bytes.
   bool BoundsCheck(uint32_t start, uint32_t end) {
-    uint32_t size = module_end - module_start;
+    size_t size = module_end - module_start;
     return start < size && end < size;
   }
